+++ tests/ha/conftest.py
@@ -1,17 +1,16 @@
 import pytest
-from common.ha.smartswitch_ha_helper import PtfTcpTestAdapter
+from pathlib import Path
 from common.ha.smartswitch_ha_io import SmartSwitchHaTrafficTest
 
 
 @pytest.fixture(scope="module")
-def copy_files(ptfhost):
-    ptfhost.copy(src="/data/tests/ha/tcp_server.py", dest='/root')
-    ptfhost.copy(src="/data/tests/ha/tcp_client.py", dest='/root')
+def deploy_files(ptfhost):
+    current_path = Path(__file__).resolve()
+    tcp_server_path = current_path.parent.parent.joinpath("common", "ha", "tcp_server.py")
+    tcp_client_path = current_path.parent.parent.joinpath("common", "ha", "tcp_client.py")
 
-
-@pytest.fixture(scope='module')
-def tcp_adapter(ptfadapter):
-    return PtfTcpTestAdapter(ptfadapter)
+    ptfhost.copy(src=str(tcp_server_path), dest='/root')
+    ptfhost.copy(src=str(tcp_client_path), dest='/root')
 
 
 @pytest.fixture(scope="module")
