+++ tests/ha/test_steady_state.py
@@ -4,7 +4,7 @@ from common.ha.smartswitch_ha_dataplane_utils import send_traffic_with_action   
 from common.ha.smartswitch_ha_helper import npu_zero, npu_one                   # noqa: F401
 
 
-def test_steady_state_active(setup_namespaces_with_routes,
+def test_steady_state_active(setup_namespaces_with_routes, deploy_files,
                              npu_zero, npu_one, send_traffic_with_action):      # noqa: F811
     """
     Verify normal operation in healthy state. Traffic lands on active DUT.
